+++ NEWS.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 0.0.3
+
+* Bug fixes for 0.0.2.
+
+## 0.0.2
+
+* Improve documentation
+* Report `electrical.batteries.*.capacity.dischargeSinceFull`.
+* Report `electrical.batteries.*.temperature` when a BMV is in
+  temperature aux mode.
+
 ## 0.0.1
 
 * Initial Release